move queries to the right place
